API improvement for Bean Widget support

+++ Modeler Core/src/org/netbeans/modeler/anchorshape/crow/OneAndOneCrowShape.java
@@ -24,17 +24,17 @@ public class OneAndOneCrowShape extends CrowAnchorShape {
         super(size);
 
         path.moveTo(size, 0.0f);
-        path.lineTo((float) size, (float) -size);
+        path.lineTo(size, -size);
         path.moveTo(size, 0.0f);
-        path.lineTo((float) size, (float) size);
+        path.lineTo(size, size);
 
         path.moveTo(0.0f, 0.0f);
-        path.lineTo((float) size, (float) 0.0f);
+        path.lineTo(size, 0.0f);
 
         path.moveTo(size / 2, 0.0f);
-        path.lineTo((float) size / 2, (float) -size);
+        path.lineTo(size / 2, -size);
         path.moveTo(size / 2, 0.0f);
-        path.lineTo((float) size / 2, (float) size);
+        path.lineTo(size / 2, size);
 
     }
 }
