+++ Modeler Core/src/org/netbeans/modeler/anchorshape/crow/ZeroOrMoreCrowShape.java
@@ -34,11 +34,11 @@ public class ZeroOrMoreCrowShape extends CrowAnchorShape {
         float centerY = 0.0f;
 
         path.moveTo(centerX, centerY);
-        path.lineTo((float) centerX - size, (float) -sin);
+        path.lineTo(centerX - size, -sin);
         path.moveTo(centerX, centerY);
-        path.lineTo((float) centerX - size, (float) sin);
+        path.lineTo(centerX - size, sin);
         path.moveTo(centerX, centerY);
-        path.lineTo((float) centerX - size, (float) 0.0f);
+        path.lineTo(centerX - size, 0.0f);
     }
 
     public void paint(Graphics2D graphics, boolean source) {
